+++ sample-tracking-status-overview-app/src/main/groovy/life/qbic/portal/sampletracking/communication/Topic.groovy
@@ -12,7 +12,9 @@ enum Topic {
     NOTIFICATION_FAILURE("Failure notification"),
     NOTIFICATION_SUCCESS("Success notification"),
     NOTIFICATION_INFO("Information notification"),
-    SAMPLE_RECEIVED_COUNT_UPDATE("The count of received samples was updated")
+    SAMPLE_RECEIVED_COUNT_UPDATE("The count of received samples was updated"),
+    SAMPLE_FAILED_QC_COUNT_UPDATE("The count of samples failing QC was updated"),
+    SAMPLE_DATA_AVAILABLE_COUNT_UPDATE("The count of available samples data was updated")
 
     @Override
     String toString() {