+++ sample-tracking-status-overview-domain/src/main/groovy/life/qbic/business/samples/count/CountSamplesOutput.groovy
@@ -1,5 +1,7 @@
 package life.qbic.business.samples.count
 
+import life.qbic.business.OutputException
+
 /**
  * <b>Output interface for the {@link CountSamples} feature</b>
  *
@@ -14,13 +16,29 @@ interface CountSamplesOutput {
      * @param reason the reason why the samples retrieval failed
      * @since 1.0.0
      */
-    void failedExecution(String reason)
+    void failedExecution(String reason) throws OutputException
 
     /**
-     * To be called after successfully counting samples for the provided code.
+     * To be called after successfully counting received samples for the provided code.
      * @param number of all samples and samples that have been received by the lab
      * @param projectCode the code of the project samples were counted for
      * @since 1.0.0
      */
-    void countedReceivedSamples(String projectCode, int allSamples, int receivedSamples)
+    void countedReceivedSamples(String projectCode, int allSamples, int receivedSamples) throws OutputException
+
+    /**
+     * To be called after successfully counting failed QC samples for the provided code.
+     * @param number of all samples and samples that have failed QC
+     * @param projectCode the code of the project samples were counted for
+     * @since 1.0.0
+     */
+    void countedFailedQcSamples(String projectCode, int allSamples, int failedQcSamples) throws OutputException
+  
+    /**
+     * To be called after successfully counting samples with available data for the provided code.
+     * @param number of all samples and samples that have available data
+     * @param projectCode the code of the project samples were counted for
+     * @since 1.0.0
+     */
+    void countedAvailableSampleData(String projectCode, int allSamples, int availableData)
 }